move metadata specific operations to its own file
