+++ app/models/concerns/advanced_search/operators/comparison_operators.rb
@@ -7,9 +7,6 @@ module AdvancedSearch
     module ComparisonOperators
       extend ActiveSupport::Concern
 
-      # Suffix convention for date-type metadata fields
-      DATE_FIELD_SUFFIX = '_date'
-
       private
 
       def condition_less_than_or_equal(scope, node, value, metadata_field:, metadata_key:)
@@ -22,53 +19,6 @@ module AdvancedSearch
         return scope.where(node.gteq(value)) unless metadata_field
 
         perform_metadata_comparison(scope, node, value, :gteq, metadata_key)
-      end
-
-      def perform_metadata_comparison(scope, node, value, comparison_method, metadata_key)
-        if date_metadata_field?(metadata_key)
-          metadata_condition_date_comparison(scope, node, value, comparison_method)
-        else
-          metadata_condition_numeric_comparison(scope, node, value, comparison_method)
-        end
-      end
-
-      def date_metadata_field?(metadata_key)
-        metadata_key.end_with?(DATE_FIELD_SUFFIX)
-      end
-
-      def metadata_condition_date_comparison(scope, node, value, comparison_method)
-        return scope.none unless valid_date_format?(value)
-
-        scope
-          .where(node.matches_regexp('^\\d{4}(-\\d{2}){0,2}$'))
-          .where(
-            Arel::Nodes::NamedFunction.new(
-              'TO_DATE', [node, Arel::Nodes::SqlLiteral.new("'YYYY-MM-DD'")]
-            ).public_send(comparison_method, value)
-          )
-      end
-
-      def metadata_condition_numeric_comparison(scope, node, value, comparison_method)
-        return scope.none unless valid_numeric_format?(value)
-
-        scope
-          .where(node.matches_regexp('^-?\\d+(\\.\\d+)?$'))
-          .where(
-            Arel::Nodes::NamedFunction.new(
-              'CAST', [node.as(Arel::Nodes::SqlLiteral.new('DOUBLE PRECISION'))]
-            ).public_send(comparison_method, value)
-          )
-      end
-
-      def valid_date_format?(value)
-        Date.iso8601(value.to_s)
-        true
-      rescue ArgumentError
-        false
-      end
-
-      def valid_numeric_format?(value)
-        value.to_s.match?(/\A-?\d+(\.\d+)?\z/)
       end
     end
   end
