+++ app/models/concerns/advanced_search/operators/set_operators.rb
@@ -20,10 +20,6 @@ module AdvancedSearch
         end
       end
 
-      def condition_in_metadata(scope, node, value)
-        scope.where(Arel::Nodes::NamedFunction.new('LOWER', [node]).in(downcase_values(value)))
-      end
-
       def condition_not_in(scope, node, value, metadata_field:, field_name:)
         # Use case-insensitive matching for metadata fields (both enum and non-enum)
         if metadata_field
@@ -34,13 +30,6 @@ module AdvancedSearch
           # Exact matching for regular fields
           scope.where(node.not_in(compact_values(value)))
         end
-      end
-
-      def condition_not_in_metadata(scope, node, value)
-        lower_function = Arel::Nodes::NamedFunction.new('LOWER', [node])
-        # Include NULL metadata values in negative set operations: NULL is not in the provided set.
-        # This maintains consistency with condition_not_equals, where "not X" includes records without the field.
-        scope.where(node.eq(nil).or(lower_function.not_in(downcase_values(value))))
       end
 
       def downcase_values(value)
